🧠Refactor(feed.repository):싱글톤 패턴 적용  - FeedListRepository를 다른 파일로 분리
