+++ src/repositories/feedSymbol.repository.ts
@@ -1,8 +1,8 @@
 import { FeedSymbol } from '../entities/feedSymbol.entity';
 import dataSource from './data-source';
-import { Pagination } from './feed.repository';
 import { Repository } from 'typeorm';
 import { CustomError } from '../utils/util';
+import { Pagination } from './feedList.repository';
 
 export class FeedSymbolRepository extends Repository<FeedSymbol> {
   constructor() {
